move misc files
