+++ lib/pages/stats_page.dart
@@ -46,71 +46,55 @@ class StatsPage extends StatelessWidget {
       10
     ];
 
-    const appBarSize = 40.0;
-    const prop1 = 0.20;
-    const prop2 = 0.18;
-    const prop3 = 0.19;
-
     return Scaffold(
-      appBar: AppBar(
-        automaticallyImplyLeading: false,
-        backgroundColor: Theme.of(context).scaffoldBackgroundColor,
-        elevation: 0,
-        toolbarHeight: appBarSize,
-        title: IconButton(
-          iconSize: appBarSize / 1.5,
-          icon: const Icon(Icons.keyboard_control),
-          color: Theme.of(context).indicatorColor,
-          onPressed: () {
-            // Perform settings action
-          },
+  body: SafeArea(
+    child: Column(
+      crossAxisAlignment: CrossAxisAlignment.center,
+      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+      children: [
+        Container(
+          height: 10,
         ),
-      ),
-      body: Column(
-          crossAxisAlignment: CrossAxisAlignment.center,
-          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-          children: [
-            Container(
-              height: 10,
-            ),
-            Expanded(
-              child: CurrentProductivityCard(
-                currentProductivity: 0.75,
-                circularProgressStroke: 20,
-                outlineStrokeWidth: 2,
-              ),
-            ),
-            Container(
-              height: 10,
-            ),
-            Expanded(
-              child: DailyPercentCompletedCard(
-                data: dailyPercentCompletedTestData,
-                barWidth: 20,
-              ),
-            ),
-            Container(
-              height: 10,
-            ),
-            Expanded(
-              child: MonthTotalTasksCompletedCard(
-                data: monthTotalCompletedTestData,
-              ),
-            ),
-            Container(
-              height: 10,
-            ),
-          ]),
-      bottomNavigationBar: BottomAppBar(
-        height: 50,
-        color: Theme.of(context).primaryColor,
-        child: IconButton(
-          icon: const Icon(Icons.close),
-          onPressed: () {
-            Navigator.pop(context);
-          },
+        Expanded(
+          child: CurrentProductivityCard(
+            currentProductivity: 0.75,
+            circularProgressStroke: 20,
+            outlineStrokeWidth: 2,
+          ),
         ),
-      ),
-    );
+        Container(
+          height: 10,
+        ),
+        Expanded(
+          child: DailyPercentCompletedCard(
+            data: dailyPercentCompletedTestData,
+            barWidth: 20,
+          ),
+        ),
+        Container(
+          height: 10,
+        ),
+        Expanded(
+          child: MonthTotalTasksCompletedCard(
+            data: monthTotalCompletedTestData,
+          ),
+        ),
+        Container(
+          height: 10,
+        ),
+      ]),
+  ),
+  bottomNavigationBar: BottomAppBar(
+    height: 50,
+    color: Theme.of(context).primaryColor,
+    child: IconButton(
+      icon: const Icon(Icons.close),
+      onPressed: () {
+        Navigator.pop(context);
+      },
+    ),
+  ),
+);
+
   }
 }
